Refactor construction of internal bounded quantified formulas
