+++ src/theory/quantifiers/quantifiers_attributes.cpp
@@ -32,7 +32,7 @@ namespace quantifiers {
 
 bool QAttributes::isStandard() const
 {
-  return !d_sygus && !d_quant_elim && !isFunDef() && !d_isInternal;
+  return !d_sygus && !d_quant_elim && !isFunDef() && !d_isQuantBounded;
 }
 
 QuantAttributes::QuantAttributes() {}
@@ -297,10 +297,10 @@ void QuantAttributes::computeQuantAttributes( Node q, QAttributes& qa ){
           qa.d_quant_elim_partial = true;
           //don't set owner, should happen naturally
         }
-        if (avar.getAttribute(InternalQuantAttribute()))
+        if (BoundedIntegers::isBoundedForallAttribute(avar))
         {
-          Trace("quant-attr") << "Attribute : internal : " << q << std::endl;
-          qa.d_isInternal = true;
+          Trace("quant-attr") << "Attribute : bounded quantifiers : " << q << std::endl;
+          qa.d_isQuantBounded = true;
         }
         if( avar.hasAttribute(QuantIdNumAttribute()) ){
           qa.d_qid_num = avar;
